+++ bloonspy/__init__.py
@@ -2,5 +2,5 @@ from .Client import *
 from .model import btd6
 from .utils import Infinity
 
-__version__ = "0.6.2"
+__version__ = "0.7.0"
 __author__ = "TheSartorsss"